Move logger
