+++ src/index.ts
@@ -18,38 +18,12 @@
 
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
-import { Logger } from './types.js'
 import { stdioToSse } from './gateways/stdioToSse.js'
 import { sseToStdio } from './gateways/sseToStdio.js'
 import { stdioToWs } from './gateways/stdioToWs.js'
 import { headers } from './lib/headers.js'
 import { corsOrigin } from './lib/corsOrigin.js'
-
-const log = (...args: any[]) => console.log('[supergateway]', ...args)
-const logStderr = (...args: any[]) => console.error('[supergateway]', ...args)
-
-const noneLogger: Logger = {
-  info: () => {},
-  error: () => {},
-}
-
-const getLogger = ({
-  logLevel,
-  outputTransport,
-}: {
-  logLevel: string
-  outputTransport: string
-}): Logger => {
-  if (logLevel === 'none') {
-    return noneLogger
-  }
-
-  if (outputTransport === 'stdio') {
-    return { info: logStderr, error: logStderr }
-  }
-
-  return { info: log, error: logStderr }
-}
+import { getLogger } from './lib/getLogger.js'
 
 async function main() {
   const argv = yargs(hideBin(process.argv))
@@ -128,18 +102,18 @@ async function main() {
   const hasStdio = Boolean(argv.stdio)
   const hasSse = Boolean(argv.sse)
 
-  if (hasStdio && hasSse) {
-    logStderr('Error: Specify only one of --stdio or --sse, not all')
-    process.exit(1)
-  } else if (!hasStdio && !hasSse) {
-    logStderr('Error: You must specify one of --stdio or --sse')
-    process.exit(1)
-  }
-
   const logger = getLogger({
     logLevel: argv.logLevel,
     outputTransport: argv.outputTransport as string,
   })
+
+  if (hasStdio && hasSse) {
+    logger.error('Error: Specify only one of --stdio or --sse, not all')
+    process.exit(1)
+  } else if (!hasStdio && !hasSse) {
+    logger.error('Error: You must specify one of --stdio or --sse')
+    process.exit(1)
+  }
 
   logger.info('Starting...')
   logger.info(
@@ -174,7 +148,7 @@ async function main() {
           healthEndpoints: argv.healthEndpoint as string[],
         })
       } else {
-        logStderr(`Error: stdio→${argv.outputTransport} not supported`)
+        logger.error(`Error: stdio→${argv.outputTransport} not supported`)
         process.exit(1)
       }
     } else if (hasSse) {
@@ -188,15 +162,15 @@ async function main() {
           }),
         })
       } else {
-        logStderr(`Error: sse→${argv.outputTransport} not supported`)
+        logger.error(`Error: sse→${argv.outputTransport} not supported`)
         process.exit(1)
       }
     } else {
-      logStderr('Error: Invalid input transport')
+      logger.error('Error: Invalid input transport')
       process.exit(1)
     }
   } catch (err) {
-    logStderr('Fatal error:', err)
+    logger.error('Fatal error:', err)
     process.exit(1)
   }
 }
